services/rbac-service/requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.0.0

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0

# Redis
redis>=5.0.0

# Authentication
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
authlib>=1.3.0
httpx>=0.26.0

# Utilities
python-multipart>=0.0.6
python-dotenv>=1.0.0
